genai-as-utilities/requirements.txt
streamlit
openai
python-dotenv
azure-storage-blob
pymupdf
streamlit-extras
azure-ai-documentintelligence==1.0.0b4
azure-ai-formrecognizer
pyodbc
azure-identity